DataEngineering/requirements.txt
antlr4-python3-runtime==4.9.3
astroid==2.15.5
colorlog==6.7.0
dill==0.3.6
exceptiongroup==1.1.1
flake8==6.0.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
iniconfig==2.0.0
isort==5.12.0
joblib==1.2.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.24.3
omegaconf==2.3.0
packaging==23.1
pandas==2.0.2
platformdirs==3.5.1
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.17.4
pytest==7.3.1
pytest-lazy-fixture==0.6.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
tzdata==2023.3
wrapt==1.15.0
